Split RowBinaryWithNamesAndTypes orchestration out of compile.ts
Keep compile.ts purely AST -> Reader (astToReader + fold helpers). Move
typeStringToReader, CompiledStream/Row, and compileRowBinaryWithNamesAndTypes
into a new rowBinaryWithNamesAndTypes.ts that ties together readHeader, the
parser, and the fold. The barrel re-exports both, so public imports are
unchanged.

Tests follow the split: compile.test.ts is now a focused astToReader unit
suite; the header/stream type corpus moves to rowBinaryWithNamesAndTypes.test.ts.
